Move `app_image` to `common/nwg_tools` [to achieve that] add `nwg_conf_inc` to nwg_tools' includedirs Remove redundant copies from `nwgdmenu`
