+++ bar/bar.cc
@@ -9,8 +9,8 @@
 
 #include <sys/time.h>
 
-#include "nwg_tools.h"
 #include "nwg_classes.h"
+#include "nwg_tools.h"
 #include "on_event.h"
 #include "bar.h"
