+++ common/nwg_tools.h
@@ -37,7 +37,6 @@ void save_json(const ns::json&, const std::string&);
 
 std::string get_output(const std::string&);
 
-/*
- * Returns x, y, width, height of focused display
- * */
+extern int image_size;
+Gtk::Image* app_image(const std::string& icon);
 Geometry display_geometry(const std::string&, Glib::RefPtr<Gdk::Display>, Glib::RefPtr<Gdk::Window>);
